requirements.txt
fastapi[standard]==0.121.1
pypdf==6.2.0
dotenv==0.9.9
psycopg2-binary==2.9.11
google-cloud-storage==3.5.0